made attraction api a single module and import to app.py
